+++ src/pyfx/view/json_lib/array/array_unexpanded_widget.py
@@ -8,10 +8,7 @@ class ArrayUnexpandedWidget(JSONWidget):
     a widget display JSON `array` type node
     """
 
-    def __init__(self,
-                 node,
-                 display_key,
-                 ):
+    def __init__(self, node, display_key):
         super().__init__(node, True, display_key)
 
     @overrides