+++ src/pyfx/view/json_lib/object/object_node.py
@@ -9,83 +9,77 @@ from ..json_composite_node import JSONCompositeNode
 
 class ObjectNode(JSONCompositeNode):
     """
-    implementation of JSON `object` type node
-    aside from fields in a JSONNode, it contains the following elements:
-    * value: dict
-    * children: dict to store correspondent node
-    * sorted_children_key_list: internal type to keep track of a sorted key list
-                                and thus keep track of the next, previous node
-                                of each child
-    * sorted_children_key_list_size: size of key
+    Implementation of JSON `object` type node.
+    Aside from fields in a JSONNode, it contains the following elements:
+
+    * children_nodes: dict to store correspondent node, note this may not have
+                      the same order as `children_key_list`, as each node of
+                      `children_nodes` is created during runtime.
+    * children_key_list: internal type to keep track of the insertion ordered
+                         key list. Thus keep track of the next, previous node
+                         of each child.
     """
 
     def __init__(self, key: str, value: dict, node_factory,
                  parent=None, display_key=True):
         super().__init__(key, value, node_factory, parent, display_key)
-        self._children = {}
-        self._sorted_children_key_list = sorted(value.keys())
-        # avoid re-calculation
-        self._sorted_children_key_list_size = len(
-            self._sorted_children_key_list
-        )
+        self._children_nodes = dict()
+        # Starting from Python 3.7, dict maintains insertion order. Thus, we
+        # don't need to sort the keys to maintain a sorted list.
+        self._children_key_list = list(value.keys())
 
     @overrides
     def collapse_all(self):
-        for key, child in self._children.items():
+        for key, child in self._children_nodes.items():
             if isinstance(child, (JSONCompositeNode, JSONCompositeEndNode)):
                 child.collapse_all()
         self.toggle_expanded()
 
     @overrides
     def has_children(self):
-        return self._sorted_children_key_list_size != 0
+        # To reduce memory cost of maintain two sets of values
+        return len(self._value) + len(self._children_nodes) != 0
 
     @overrides
     def get_first_child(self):
         if not self.has_children():
             return None
-        return self._get_child_node(self._sorted_children_key_list[0])
+        return self._get_child_node(self._children_key_list[0])
 
     @overrides
     def get_last_child(self):
         if not self.has_children():
             return None
-        return self._get_child_node(
-            self._sorted_children_key_list[
-                self._sorted_children_key_list_size - 1
-            ]
-        )
+        return self._get_child_node(self._children_key_list[-1])
 
     @overrides
     def prev_child(self, key):
-        index = self._sorted_children_key_list.index(key)
+        index = self._children_key_list.index(key)
         if index == 0:
             return None
-        return self._get_child_node(self._sorted_children_key_list[index - 1])
+        return self._get_child_node(self._children_key_list[index - 1])
 
     @overrides
     def next_child(self, key):
-        index = self._sorted_children_key_list.index(key)
-        if index == self._sorted_children_key_list_size - 1:
+        index = self._children_key_list.index(key)
+        if index == len(self._children_key_list) - 1:
             return None
-        return self._get_child_node(self._sorted_children_key_list[index + 1])
+        return self._get_child_node(self._children_key_list[index + 1])
 
     def _get_child_node(self, key):
         if not self.has_children():
             return None
-        elif key not in self._children:
-            self._children[key] = self._load_child_node(key)
-        return self._children[key]
+        elif key not in self._children_nodes:
+            self._children_nodes[key] = self._load_child_node(key)
+            # Remove the according value to reduce memory overhead,
+            # as now it's value is stored in `children_nodes`.
+            del self._value[key]
+        return self._children_nodes[key]
 
     def _load_child_node(self, key):
         value = self.get_value()[key]
-        return self._node_factory.create_node(
-            key, value, parent=self, display_key=True
-        )
-
-    # ====================================================================== #
-    # ui                                                                     #
-    # ====================================================================== #
+        return self._node_factory.create_node(key, value, parent=self,
+                                              display_key=True)
 
     @overrides
     def load_unexpanded_widget(self):